ControlAllocation: update normalization scale for only if matrix updated is forced
The forced flag is used to distinguish when the matrix is updated because the user changes the
configuration (only enabled when disarmed), and when the matrix is updated because the
geometry changes due to motors being tilted.
In the future another reason for updating the matrix can be detection an actuator failure.

Signed-off-by: Silvan Fuhrer <silvan@auterion.com>